test: improve e2e workspace
