+++ tests/integration/Cargo.toml
@@ -20,8 +20,9 @@ serde_json = { version = "1.0", default-features = false }
 sqlx = { version = "0.6.3", features = ["postgres", "runtime-async-std-rustls", "decimal", "chrono", "uuid"] }
 rust_decimal = { version = "1.36" }
 chrono = { version = "0.4" }
+once_cell = "1.20.2"
 uuid = { version = "1.11", features = ["v4"] }
 
 fluvio = { workspace = true }
 fluvio-future = { workspace = true }
-fluvio-model-sql = { path = "../fluvio-model-sql"}
+fluvio-model-sql = { workspace = true }
