+++ tests/integration/configs/test_postgres_all_data_types.yaml
