+++ tests/integration/configs/test_postgres_consumer_offsets.yaml
