+++ tests/integration/configs/test_postgres_with_json_sql_transformations.yaml
